Recipes/requirements.txt
anyio==3.7.0
asgiref==3.6.0
astroid==2.15.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
dill==0.3.6
Django==3.2.18
django-appconf==1.0.5
django-imagekit==4.1.0
django-phonenumber-field==7.1.0
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
isort==5.11.5
lazy-object-proxy==1.9.0
mccabe==0.7.0
phonenumbers==8.13.13
pilkit==2.0
Pillow==9.5.0
platformdirs==3.2.0
pylint==2.17.2
python-dotenv==0.21.1
python-telegram-bot==20.3
pytz==2023.3
requests==2.28.2
six==1.16.0
sniffio==1.3.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.7
typed-ast==1.5.4
typing-extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
